@@ -5,14 +5,16 @@
   "keywords": [],
   "main": "src/index.js",
   "dependencies": {
-    "html2canvas": "1.4.0",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
-    "react-scripts": "4.0.0"
+    "draft-js": "^0.11.7",
+    "html2canvas": "1.4.1",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
+    "react-draft-wysiwyg": "^1.15.0",
+    "react-scripts": "5.0.1"
   },
   "devDependencies": {
-    "@babel/runtime": "7.13.8",
-    "typescript": "4.1.3"
+    "@babel/runtime": "7.20.7",
+    "typescript": "4.9.4"
   },
   "scripts": {
     "start": "react-scripts start",
