@@ -16,7 +16,11 @@ const exportAsImage = async (element, imageFileName) => {
   html.style.width = htmlWidth + "px";
   body.style.width = bodyWidth + "px";
 
-  const canvas = await html2canvas(element);
+  const canvas = await html2canvas(element, {
+    backgroundColor: null,
+    allowTaint: true,
+    useCORS: true
+  });
   const image = canvas.toDataURL("image/png", 1.0);
   downloadImage(image, imageFileName);
   html.style.width = null;
